requirements.txt
numpy
pandas
psycopg2-binary
requests
tqdm
unidecode
lxml
pyyaml
jsonslicer